Format the repository with Runic
The Runic format check fails on master for 102 files (formatting drift
predating the SciML workflows onboarding). This is a mechanical
`runic -i .` pass with Runic v1.10.0, matching CI.

Agent-Harness: Devin CLI 3000.10.21
Agent-Model: SWE-2 Max
Agent-Session: local session, /home/crackauc/.local/share/devin/cli/sessions.db

+++ examples/ivp_2d_shear_flow/shear_flow.jl
@@ -27,25 +27,25 @@ logger = Logging.current_logger()
 # Parameters
 Lx, Lz = 1, 2
 Nx, Nz = 128, 256
-Reynolds = 5e4
+Reynolds = 5.0e4
 Schmidt = 1
-dealias = 3/2
+dealias = 3 / 2
 stop_sim_time = 20
 timestepper = RK222
-max_timestep = 1e-2
+max_timestep = 1.0e-2
 dtype = Float64
 
 # Bases
 coords = CartesianCoordinates("x", "z")
-dist = Distributor(coords; dtype=dtype)
-xbasis = RealFourier(coords["x"], Nx; bounds=(0, Lx), dealias=dealias)
-zbasis = RealFourier(coords["z"], Nz; bounds=(-Lz/2, Lz/2), dealias=dealias)
+dist = Distributor(coords; dtype = dtype)
+xbasis = RealFourier(coords["x"], Nx; bounds = (0, Lx), dealias = dealias)
+zbasis = RealFourier(coords["z"], Nz; bounds = (-Lz / 2, Lz / 2), dealias = dealias)
 
 # Fields
-p = Field(dist; name="p", bases=(xbasis, zbasis))
-s = Field(dist; name="s", bases=(xbasis, zbasis))
-u = VectorField(dist, coords; name="u", bases=(xbasis, zbasis))
-tau_p = Field(dist; name="tau_p")
+p = Field(dist; name = "p", bases = (xbasis, zbasis))
+s = Field(dist; name = "s", bases = (xbasis, zbasis))
+u = VectorField(dist, coords; name = "u", bases = (xbasis, zbasis))
+tau_p = Field(dist; name = "tau_p")
 
 # Substitutions
 nu = 1 / Reynolds
@@ -54,7 +54,7 @@ x, z = local_grids(dist, xbasis, zbasis)
 ex, ez = unit_vector_fields(coords, dist)
 
 # Problem
-problem = IVP([u, s, p, tau_p]; namespace=@locals)
+problem = IVP([u, s, p, tau_p]; namespace = @locals)
 add_equation!(problem, "dt(u) + grad(p) - nu*lap(u) = - u@grad(u)")
 add_equation!(problem, "dt(s) - D*lap(s) = - u@grad(s)")
 add_equation!(problem, "div(u) + tau_p = 0")
@@ -66,27 +66,29 @@ solver.stop_sim_time = stop_sim_time
 
 # Initial conditions
 # Background shear
-u["g"][1] = @. 1/2 + 1/2 * (tanh((z - 0.5) / 0.1) - tanh((z + 0.5) / 0.1))
+u["g"][1] = @. 1 / 2 + 1 / 2 * (tanh((z - 0.5) / 0.1) - tanh((z + 0.5) / 0.1))
 # Match tracer to shear
 s["g"] = u["g"][1]
 # Add small vertical velocity perturbations localized to the shear layers
-u["g"][2] .+= @. 0.1 * sin(2*pi*x/Lx) * exp(-(z - 0.5)^2 / 0.01)
-u["g"][2] .+= @. 0.1 * sin(2*pi*x/Lx) * exp(-(z + 0.5)^2 / 0.01)
+u["g"][2] .+= @. 0.1 * sin(2 * pi * x / Lx) * exp(-(z - 0.5)^2 / 0.01)
+u["g"][2] .+= @. 0.1 * sin(2 * pi * x / Lx) * exp(-(z + 0.5)^2 / 0.01)
 
 # Analysis
-snapshots = add_file_handler(solver.evaluator, "snapshots"; sim_dt=0.1, max_writes=10)
-add_task!(snapshots, s; name="tracer")
-add_task!(snapshots, p; name="pressure")
-add_task!(snapshots, -divergence(skew(u)); name="vorticity")
+snapshots = add_file_handler(solver.evaluator, "snapshots"; sim_dt = 0.1, max_writes = 10)
+add_task!(snapshots, s; name = "tracer")
+add_task!(snapshots, p; name = "pressure")
+add_task!(snapshots, -divergence(skew(u)); name = "vorticity")
 
 # CFL
-cfl = CFL(solver; initial_dt=max_timestep, cadence=10, safety=0.2, threshold=0.1,
-          max_change=1.5, min_change=0.5, max_dt=max_timestep)
+cfl = CFL(
+    solver; initial_dt = max_timestep, cadence = 10, safety = 0.2, threshold = 0.1,
+    max_change = 1.5, min_change = 0.5, max_dt = max_timestep
+)
 add_velocity!(cfl, u)
 
 # Flow properties
-flow = GlobalFlowProperty(solver; cadence=10)
-add_property!(flow, DotProduct(u, ez)^2; name="w2")
+flow = GlobalFlowProperty(solver; cadence = 10)
+add_property!(flow, DotProduct(u, ez)^2; name = "w2")
 
 # Main loop
 try
