+++ examples/ivp_disk_libration/libration.jl
@@ -31,23 +31,23 @@ logger = Logging.current_logger()
 Nphi, Nr = 32, 128
 Ekman = 1 / 2 / 20^2
 Ro = 40
-dealias = 3/2
+dealias = 3 / 2
 stop_sim_time = 50
 timestepper = SBDF2
-timestep = 1e-3
+timestep = 1.0e-3
 dtype = Float64
 
 # Bases
 coords = PolarCoordinates("phi", "r")
-dist = Distributor(coords; dtype=dtype)
-disk = DiskBasis(coords; shape=(Nphi, Nr), radius=1, dealias=dealias, dtype=dtype)
+dist = Distributor(coords; dtype = dtype)
+disk = DiskBasis(coords; shape = (Nphi, Nr), radius = 1, dealias = dealias, dtype = dtype)
 disk_edge = edge(disk)
 
 # Fields
-u = VectorField(dist, coords; name="u", bases=(disk,))
-p = Field(dist; name="p", bases=(disk,))
-tau_u = VectorField(dist, coords; name="tau_u", bases=(disk_edge,))
-tau_p = Field(dist; name="tau_p")
+u = VectorField(dist, coords; name = "u", bases = (disk,))
+p = Field(dist; name = "p", bases = (disk,))
+tau_u = VectorField(dist, coords; name = "tau_u", bases = (disk_edge,))
+tau_p = Field(dist; name = "tau_p")
 
 # Substitutions
 phi, r = local_grids(dist, disk)
@@ -55,15 +55,15 @@ nu = Ekman
 lift = A -> Lift(A, disk, -1)
 
 # Background librating flow
-u0_real = VectorField(dist, coords; bases=(disk,))
-u0_imag = VectorField(dist, coords; bases=(disk,))
+u0_real = VectorField(dist, coords; bases = (disk,))
+u0_imag = VectorField(dist, coords; bases = (disk,))
 u0_real["g"][1] = @. Ro * real(besselj(1, (1 - 1im) * r / sqrt(2 * Ekman)) / besselj(1, (1 - 1im) / sqrt(2 * Ekman)))
 u0_imag["g"][1] = @. Ro * imag(besselj(1, (1 - 1im) * r / sqrt(2 * Ekman)) / besselj(1, (1 - 1im) / sqrt(2 * Ekman)))
 t = Field(dist)
 u0 = cos(t) * u0_real - sin(t) * u0_imag
 
 # Problem
-problem = IVP([p, u, tau_u, tau_p]; time=t, namespace=@locals)
+problem = IVP([p, u, tau_u, tau_p]; time = t, namespace = @locals)
 add_equation!(problem, "div(u) + tau_p = 0")
 add_equation!(problem, "dt(u) - nu*lap(u) + grad(p) + lift(tau_u) = - u@grad(u0) - u0@grad(u)")
 add_equation!(problem, "u(r=1) = 0")
@@ -74,18 +74,18 @@ solver = build_solver(problem, timestepper)
 solver.stop_sim_time = stop_sim_time
 
 # Initial conditions
-fill_random!(u, "g"; seed=42, distribution="standard_normal")  # Random noise
-low_pass_filter!(u; scales=0.25)  # Keep only lower fourth of the modes
+fill_random!(u, "g"; seed = 42, distribution = "standard_normal")  # Random noise
+low_pass_filter!(u; scales = 0.25)  # Keep only lower fourth of the modes
 
 # Analysis
-snapshots = add_file_handler(solver.evaluator, "snapshots"; sim_dt=0.1, max_writes=20)
-add_task!(snapshots, -divergence(skew(u)); scales=(4, 1), name="vorticity")
-scalars = add_file_handler(solver.evaluator, "scalars"; sim_dt=0.01)
-add_task!(scalars, integrate(0.5 * DotProduct(u, u)); name="KE")
+snapshots = add_file_handler(solver.evaluator, "snapshots"; sim_dt = 0.1, max_writes = 20)
+add_task!(snapshots, -divergence(skew(u)); scales = (4, 1), name = "vorticity")
+scalars = add_file_handler(solver.evaluator, "scalars"; sim_dt = 0.01)
+add_task!(scalars, integrate(0.5 * DotProduct(u, u)); name = "KE")
 
 # Flow properties
-flow = GlobalFlowProperty(solver; cadence=100)
-add_property!(flow, DotProduct(u, u); name="u2")
+flow = GlobalFlowProperty(solver; cadence = 100)
+add_property!(flow, DotProduct(u, u); name = "u2")
 
 # Main loop
 try
